Refactored data models to support `double`/`num` types for nutrition tracking, enhanced the onboarding and authentication flow, and improved UI consistency across the dashboard and settings.
### Data & State Management
- **Type Refactoring:** Migrated nutrient fields (calories, protein, carbs, fats, etc.) from `int` to `double` or `num` across `FoodLog`, `Nutrition`, `NutritionGoals`, and `NutritionProgress` models to support precise calculations.
- **Model Updates:**
    - Updated `User` model to use `uid` consistently and improved JSON serialization for `UserProfile` and `UserGoal`.
    - Introduced `ObstacleType` enum to replace raw strings for maintenance strategies.
    - Refactored `FoodLog.fromFoodFact` and other factory methods to handle numeric casting safely.
- **Firestore Service:**
    - Optimized `updateFoodEntry` and added `updateExerciseEntry` using atomic batch updates and numeric deltas.
    - Improved `saveFullUser` logic to prevent overwriting specific profile fields (like provider or name) with default values during anonymous-to-google linking.
    - Enhanced weight logging to initialize history documents if they do not exist.

### Onboarding & Authentication
- **Account Linking:** Implemented `linkGoogleAccount` in `AuthService` to allow anonymous users to link their progress to a Google account without data loss.
- **Enhanced Step 19:** Added logic to automatically skip the login/subscription screen if a non-anonymous user session is detected.
- **Goal Editing:** Integrated `EditModal` for real-time adjustment of nutrition targets during the final onboarding summary.
- **Bug Fixes:** Fixed state persistence issues in `Step 9.5` (obstacles) and `Step 14/15` (settings toggles) by correctly utilizing `userNotifier.updateLocal`.

### UI & UX Enhancements
- **New Views:**
    - Added `BmiInfoView` providing detailed information about BMI categories and health risks.
    - Implemented `EditGoalsView` for comprehensive adjustment of macro and micro nutrient targets.
    - Created `EditExercisePage` for modifying logged activity intensity and duration.
- **Dashboard & Home:**
    - Optimized `HomeBody` and `CarouselView` using `RepaintBoundary` and Riverpod's `.select` to minimize unnecessary rebuilds.
    - Improved `ActivityCard` and `CarouselItemCalories` to handle rounded display values for non-integer inputs.
    - Added a water settings sheet to `CarouselActivity` for future configuration of serving sizes.
- **Visual Polish:**
    - Updated `ProgressBmiCard` with a more flexible layout and support for navigation to info pages.
    - Standardized `CustomAppBar` usage across several screens (`FoodDatabase`, `SelectedFoodPage`, `EditGoals`).
    - Fixed text visibility issues in `UnitToggle` and updated graph colors to use `colorScheme.onPrimary`.

### Settings & Profile
- **Account Management:** Added an account deletion flow with a confirmation dialog that wipes both Firestore data and the Firebase Auth account.
- **Weight History:** Integrated real-time Firestore data into the `WeightHistoryView`, displaying entries sorted chronologically.
- **Linking UI:** Added a prominent "Back up your data" card for anonymous users in the settings menu.

+++ lib/models/nutrition_model.dart
@@ -3,15 +3,15 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 
 abstract class Nutrition {
-  final int calories;
-  final int protein;
-  final int carbs;
-  final int fats;
-  final int fiber;
-  final int sugar;
-  final int sodium;
-  final int water;
-  final int steps;
+  final double calories;
+  final double protein;
+  final double carbs;
+  final double fats;
+  final double fiber;
+  final double sugar;
+  final double sodium;
+  final double water;
+  final double steps;
 
   const Nutrition({
     this.calories = 0,
@@ -27,8 +27,8 @@ abstract class Nutrition {
 }
 
 class NutritionGoals extends Nutrition {
-  final int weightGoal;
-  final int rollover;
+  final double weightGoal;
+  final double rollover;
 
   const NutritionGoals({
     super.calories,
@@ -46,7 +46,7 @@ class NutritionGoals extends Nutrition {
 
   factory NutritionGoals.fromJson(Map<String, dynamic> map) {
     // Use a nullable return type here
-    int? val(String key) => (map[key] as num?)?.toInt();
+    double? val(String key) => (map[key] as num?)?.toDouble();
 
     return NutritionGoals(
       calories: val('calorieGoal') ?? 0,
@@ -58,7 +58,7 @@ class NutritionGoals extends Nutrition {
       sodium: val('sodiumGoal') ?? 0,
       water: val('waterGoal') ?? 64,
       steps: val('stepsGoal') ?? 10000,
-      weightGoal: val('weightGoal') ?? 0 ,
+      weightGoal: val('weightGoal') ?? 0,
       rollover: val('rollover') ?? 0,
     );
   }
@@ -80,17 +80,17 @@ class NutritionGoals extends Nutrition {
   static const empty = NutritionGoals();
 
   NutritionGoals copyWith({
-    int? calories,
-    int? protein,
-    int? carbs,
-    int? fats,
-    int? fiber,
-    int? sugar,
-    int? sodium,
-    int? water,
-    int? steps,
-    int? targetWeight,
-    int? rollover,
+    double? calories,
+    double? protein,
+    double? carbs,
+    double? fats,
+    double? fiber,
+    double? sugar,
+    double? sodium,
+    double? water,
+    double? steps,
+    double? targetWeight,
+    double? rollover,
   }) {
     return NutritionGoals(
       calories: calories ?? this.calories,
@@ -144,7 +144,7 @@ class NutritionLog extends Nutrition {
   factory NutritionLog.fromSnapshot(DateTime date, Map<String, dynamic> data) {
     final dp = data['dailyProgress'] as Map<String, dynamic>? ?? {};
 
-    int val(Map m, String k) => (m[k] as num?)?.toInt() ?? 0;
+    double val(Map m, String k) => (m[k] as num?)?.toDouble() ?? 0;
 
     return NutritionLog(
       date: date,
@@ -162,7 +162,7 @@ class NutritionLog extends Nutrition {
 }
 
 class NutritionProgress extends Nutrition {
-  final int caloriesBurned;
+  final double caloriesBurned;
 
   const NutritionProgress({
     super.calories, super.protein, super.carbs, super.fats,
@@ -188,7 +188,7 @@ class NutritionProgress extends Nutrition {
   factory NutritionProgress.fromDailyLog(Map<String, dynamic> data) {
     final dp = data['dailyProgress'] as Map<String, dynamic>? ?? {};
 
-    int val(Map m, String k) => (m[k] as num?)?.toInt() ?? 0;
+    double val(Map m, String k) => (m[k] as num?)?.toDouble() ?? 0;
 
     return NutritionProgress(
       calories: val(dp, 'caloriesEaten'),
